bplip. veliko dela še

+++ sw/programs/keyboard/keypad.h
@@ -3,4 +3,4 @@
 #define numRow 4
 #define numCol 4
 
-uint32_t getKeyPress(uint32_t rows[], uint32_t cols[]);
+uint16_t getKeyPress(uint16_t rows[], uint16_t cols[]);
